+++ src/main/java/dinkplugin/notifiers/DeathNotifier.java
@@ -25,6 +25,7 @@ import net.runelite.api.Prayer;
 import net.runelite.api.Varbits;
 import net.runelite.api.events.ActorDeath;
 import net.runelite.api.events.InteractingChanged;
+import net.runelite.api.events.ScriptPreFired;
 import net.runelite.client.game.ItemManager;
 import net.runelite.client.game.NPCManager;
 import org.apache.commons.lang3.ArrayUtils;
@@ -57,6 +58,13 @@ public class DeathNotifier extends BaseNotifier {
     private static final String ATTACK_OPTION = "Attack";
 
     private static final String TOA_DEATH_MSG = "You failed to survive the Tombs of Amascut";
+
+    private static final String TOB_DEATH_MSG = "Your party has failed";
+
+    /**
+     * @see <a href="https://github.com/Joshua-F/cs2-scripts/blob/master/scripts/%5Bclientscript,tob_hud_portal%5D.cs2">CS2 Reference</a>
+     */
+    private static final int TOB_HUB_PORTAL_SCRIPT = 2307;
 
     /**
      * Checks whether the actor is alive and interacting with the specified player.
@@ -117,11 +125,24 @@ public class DeathNotifier extends BaseNotifier {
     }
 
     public void onGameMessage(String message) {
-        if (config.deathIgnoreSafe() && !config.deathSafeExceptions().contains(ExceptionalDeath.TOA) &&
-            !Utils.getAccountType(client).isHardcore() && message.contains(TOA_DEATH_MSG) && isEnabled()) {
+        if (shouldNotifyExceptionalDangerousDeath(ExceptionalDeath.TOA) && message.contains(TOA_DEATH_MSG)) {
             // https://github.com/pajlads/DinkPlugin/issues/316
             // though, hardcore (group) ironmen just use the normal ActorDeath trigger for TOA
             handleNotify(Danger.DANGEROUS);
+        }
+    }
+
+    public void onScript(ScriptPreFired event) {
+        if (event.getScriptId() == TOB_HUB_PORTAL_SCRIPT && event.getScriptEvent() != null &&
+            shouldNotifyExceptionalDangerousDeath(ExceptionalDeath.TOB)) {
+            Object[] args = event.getScriptEvent().getArguments();
+            if (args != null && args.length > 1) {
+                Object text = args[1];
+                if (text instanceof String && ((String) text).contains(TOB_DEATH_MSG)) {
+                    // https://oldschool.runescape.wiki/w/Theatre_of_Blood#Death_within_the_Theatre
+                    handleNotify(Danger.DANGEROUS);
+                }
+            }
         }
     }
 
@@ -217,6 +238,26 @@ public class DeathNotifier extends BaseNotifier {
             builder.replacement("%NPC%", Replacements.ofWiki(killer));
         }
         return builder.build();
+    }
+
+    private boolean shouldNotifyExceptionalDangerousDeath(ExceptionalDeath death) {
+        if (!config.deathIgnoreSafe()) {
+            // safe death notifications are enabled => we already notified
+            return false;
+        }
+
+        if (config.deathSafeExceptions().contains(death)) {
+            // safe deaths are ignored, but this death is exceptional => we already notified
+            return false;
+        }
+
+        if (Utils.getAccountType(client).isHardcore() && death != ExceptionalDeath.FIGHT_CAVE) {
+            // the PvM death is actually dangerous since hardcore => we already notified
+            return false;
+        }
+
+        // notifier must be enabled to dink when the actually dangerous death occurs
+        return isEnabled();
     }
 
     /**